pyproject.toml
[project]
name = "mcp-knowledge-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "content-extraction>=0.3.1",
    "django>=5.2.5",
    "django-ensuresuperuser>=0.1.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "google-cloud-storage>=3.3.0",
    "hypercorn>=0.17.3",
    "uuid-utils>=0.11.0",
    "psycopg2-binary>=2.9.10",
    "pycasbin>=2.2.0",
    "django-content-access-control>=0.3.0",
    "casbin>=1.43.0",
    "google-genai>=1.34.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-django>=4.11.1",
    "ruff>=0.12.11",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "backend.settings"
